Allow empty daily folder files
Fixed an issue where the content of template file could not be empty. See https://github.com/NemoAndrea/obsidian-daily-named-folder/issues/6. This lead to opaque issues to users where the ribbon icon was not working, see: https://github.com/NemoAndrea/obsidian-daily-named-folder/issues/2. Big thanks to @bvn13 for some detectivework and for bringing the source of the problem to my attention. To fix it, I have added a toggle button to settings that users can disable to skip the template file. I think this is a better UX solution than allowing an empty filepath string to mean 'no template'. Changed the error message to be more informative. And finally rearranged the sequence of new file creation so that the error for misconfigured template will pop up _before_ the user fills out the name for the daily folder.